add hamburger menu to app
