+++ src/css/header.css
@@ -39,11 +39,6 @@ button {
     padding: 20px; 
 } 
 
-.hamburgerMenu {
-    height: 100%;
-    background-color: aqua;
-}
-
 /* .burgerContainer {
     display: none;
 } */
